requirements.txt
Flask==2.0.1
Flask-WTF==1.2.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.31.0
Werkzeug==2.0.2
WTForms==3.1.2
python-dotenv==1.0.1